Moves database adapter files into common directory
