+++ src/pages/api/accounts.ts
@@ -1,8 +1,8 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import type { AccountsGetRequest, CountryCode, InstitutionsGetByIdRequest } from 'plaid';
 
-import * as db from '@/lib/database/Adapter';
-import * as messages from '@/lib/Messages';
+import * as db from '@/common/utils/database/Adapter';
+import * as messages from '@/common/Messages';
 import plaidClient from '@/lib/PlaidApiClient';
 
 export interface AccountData {
