+++ src/scripts/CreateDatabase.ts
@@ -1,6 +1,6 @@
 // this file must be executed from project root (email-summary/)
 
-import { createTables } from '../lib/database/AppDAO';
+import { createTables } from '../common/utils/database/AppDAO';
 
 (async () => {
   await createTables();
